requirements.txt
# Plutus - Wealth Building Coach Brain Dependencies
# =================================================
# These are the additional dependencies needed for Plutus when integrated with Wealthify
# Note: Wealthify already has FastAPI, SQLAlchemy, Redis, etc.

# Core AI Dependencies
anthropic>=0.60.0                # Claude API client (latest version)
python-dotenv>=1.0.0            # Environment variable loading

# Optional: LangGraph (if using advanced orchestration)
# langgraph>=0.2.0               # Uncomment if using LangGraph workflows
# langchain-anthropic>=0.2.0     # Uncomment if using LangChain integration

# Data Processing (if not already in Wealthify)
# pandas>=2.0.0                  # Uncomment if Wealthify doesn't have pandas
# numpy>=1.24.0                  # Uncomment if Wealthify doesn't have numpy

# Utilities
httpx>=0.27.0                    # HTTP client for API calls
tenacity>=9.0.0                  # Retry logic for API resilience

# Development/Testing (for Plutus development only)
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Note: The following are likely already in Wealthify's requirements:
# - fastapi
# - uvicorn  
# - pydantic
# - sqlalchemy
# - redis
# - structlog or similar logging